Changes file names and routes
A. Renamed files
	1. tentang-kami.js 	-> about-us.js
	2. tentang.js 		-> about.js
	3. tentang.scss 	-> about.scss

B. Update link routes in files
	1. change navigation-bar.js	(navList)
	2. story-detail.js		(back button href)
	3. routes.js			(change routes to explore and about)
	4. main.scss			(change import to about)
