+++ src/scripts/components/navigation-bar.js
@@ -18,9 +18,9 @@ class NavigationBar extends HTMLElement {
       <nav id="navList" class="nav-list">
         <ul>
           <li><a href="#">Beranda</a></li>
-          <li><a href="#/explore-page">Cerita</a></li>
+          <li><a href="#/explore">Cerita</a></li>
           <li><a href="#/favorite">Favorite</a></li>
-          <li><a href="#/tentang">Tentang Kami</a></li>
+          <li><a href="#/about">Tentang Kami</a></li>
         </ul>
       </nav>
     `;
