+++ src/styles/about.scss
@@ -1,4 +1,4 @@
-.tentang {
+.about {
   padding: 40px 20px;
   text-align: center;
   background-color: #f9f9f9;
