+++ src/styles/main.scss
@@ -100,5 +100,5 @@ textarea {
 @import "genre";
 @import "detail";
 @import "footer";
-@import "tentang";
+@import "about";
 @import "explore";
